creature: move Creature_Die to trx
