+++ src/libtrx/game/creature/common.c
@@ -2,6 +2,7 @@
 #include "game/lara/common.h"
 #include "game/los.h"
 #include "game/objects/vars.h"
+#include "game/pathing/lot.h"
 #include "game/random.h"
 #include "game/rooms.h"
 #include "log.h"
@@ -27,7 +28,9 @@ static bool M_SwitchToLand(
     int16_t item_num, const int32_t *wh, const HYBRID_INFO *info);
 static bool M_TestSwitchOrKill(int16_t item_num, GAME_OBJECT_ID target_id);
 
-#if TR_VERSION == 2
+#if TR_VERSION == 1
+extern void Carrier_TestItemDrops(int16_t item_num);
+#elif TR_VERSION == 2
 extern void Creature_GetBaddieTarget(int16_t item_num, bool goody);
 #endif
 
@@ -808,6 +811,66 @@ bool Creature_Animate(
         Item_NewRoom(item_num, room_num);
     }
     return true;
+}
+
+void Creature_Die(const int16_t item_num, const bool explode)
+{
+    ITEM *const item = Item_Get(item_num);
+
+    switch (item->object_id) {
+#if TR_VERSION == 2
+    case O_DRAGON_FRONT:
+        item->hit_points = 0;
+        return;
+
+    case O_SKIDOO_DRIVER:
+        if (explode) {
+            Item_Explode(item_num, -1, 0);
+        }
+        item->hit_points = DONT_TARGET;
+        const int16_t vehicle_item_num = (int16_t)(intptr_t)item->data;
+        ITEM *const vehicle_item = Item_Get(vehicle_item_num);
+        vehicle_item->hit_points = 0;
+        return;
+#endif
+
+    default:
+        break;
+    }
+
+    item->collidable = 0;
+    item->hit_points = DONT_TARGET;
+    if (explode) {
+        Item_Explode(item_num, -1, 0);
+        Item_Kill(item_num);
+    } else {
+        Item_RemoveActive(item_num);
+    }
+
+    const OBJECT *const obj = Object_Get(item->object_id);
+    if (obj->intelligent) {
+        LOT_DisableBaddieAI(item_num);
+    }
+    item->flags |= IF_ONE_SHOT;
+
+#if TR_VERSION >= 2
+    if (item->killed) {
+        item->next_active = Item_GetPrevActive();
+        Item_SetPrevActive(item_num);
+    }
+
+    if (obj->intelligent) {
+        int16_t pickup_num = item->carried_item;
+        while (pickup_num != NO_ITEM) {
+            ITEM *const pickup = Item_Get(pickup_num);
+            pickup->pos = item->pos;
+            Item_NewRoom(pickup_num, item->room_num);
+            pickup_num = pickup->carried_item;
+        }
+    }
+#else
+    Carrier_TestItemDrops(item_num);
+#endif
 }
 
 int16_t Creature_Effect(
