+++ src/libtrx/include/libtrx/game/pathing/lot.h
@@ -1,3 +1,4 @@
 #pragma once
 
 extern bool LOT_EnableBaddieAI(int16_t item_num, bool always);
+extern void LOT_DisableBaddieAI(int16_t item_num);
