+++ src/tr1/game/lot.h
@@ -7,7 +7,6 @@
 #include <stdint.h>
 
 void LOT_InitialiseArray(void);
-void LOT_DisableBaddieAI(int16_t item_num);
 void LOT_InitialiseSlot(int16_t item_num, int32_t slot);
 void LOT_CreateZone(ITEM *item);
 void LOT_InitialiseLOT(LOT_INFO *LOT);
