+++ src/tr2/game/lot.h
@@ -5,7 +5,6 @@
 #include <libtrx/game/pathing/lot.h>
 
 void LOT_InitialiseArray(void);
-void LOT_DisableBaddieAI(int16_t item_num);
 void LOT_InitialiseSlot(int16_t item_num, int32_t slot);
 void LOT_CreateZone(ITEM *item);
 void LOT_ClearLOT(LOT_INFO *LOT);
